+++ src/.ipynb_checkpoints/functions-checkpoint.py
@@ -1,26 +1,10 @@
-import multiprocessing
-import time  # Ensure time module is imported
+import random
+import string
 
-# Function to wrap each task in a process
-def process_function(func, *args):
-    return func(*args)
+# Function to generate 1000 random characters and join them into a string
+def generate_random_chars():
+    return ''.join(random.choices(string.ascii_letters, k=1000))
 
-# Function to run the processes task
-def run_processes():
-    from src.functions import generate_random_chars, generate_random_sum  # Import inside function
-
-    # Start processes for both tasks
-    process1 = multiprocessing.Process(target=process_function, args=(generate_random_chars,))
-    process2 = multiprocessing.Process(target=process_function, args=(generate_random_sum, 2, 3))  # Pass args correctly
-    
-    # Run the processes
-    start_time = time.time()  # Measure start time
-    process1.start()
-    process2.start()
-
-    # Wait for both processes to finish
-    process1.join()
-    process2.join()
-    
-    end_time = time.time()  # Measure end time
-    print(f"Time to execute using processes: {end_time - start_time} seconds.")
+# Function to generate 1000 random numbers and add them
+def generate_random_sum():
+    return sum(random.choices(range(1, 101), k=1000))
